+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- DescribePCR subcommand to get PCR value by its index
 - This CHANGELOG file
 
 ### Changed